requirements.txt
python-telegram-bot[job-queue]>=21.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
SQLAlchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.1
arq>=0.25.0
sentry-sdk>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0
boto3>=1.34.0
alembic>=1.14.0
pdf2image
openpyxl>=3.1.5
pypdf>=6.0.0
